apps/drones/api/serializers.py
from rest_framework import serializers
from apps.drones.models import Drone
from apps.drones.models import Medication

class DroneSerializer(serializers.ModelSerializer):
    class Meta:
        model = Drone
        fields = '__all__'



class MedicationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Medication
        fields = '__all__'