requirements.txt
django~=4.2.1
djangorestframework~=3.14.0
apscheduler~=3.10.1
django-simple-history~=3.3.0